the assignment is ready

+++ helperFunctions.h
@@ -7,7 +7,7 @@
 #include <time.h>
 
 
-int isPrime(int);
+int isPrime(size_t);
 int AOTask1(void* pao);
 int AOTask2(void* pao);
 int AOTask3(void* pao);
